% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aho_corasick_Rcpp}
\alias{aho_corasick_Rcpp}
\title{Aho-Corasick multiple pattern matching algorithm using C++}
\usage{
aho_corasick_Rcpp(text, patterns)
}
\arguments{
\item{text}{A character string representing the main text.}

\item{patterns}{A character vector containing multiple patterns to search for.}
}
\value{
A list of integers indicating the starting positions of each match (1-based index).
}
\description{
Searches for all occurrences of multiple patterns in a given text using the Aho-Corasick algorithm.
}
