% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_pattern_matching.R
\name{naive_pattern_matching}
\alias{naive_pattern_matching}
\title{Naive Pattern Matching Algorithm}
\usage{
naive_pattern_matching(texte, motif)
}
\arguments{
\item{texte}{A character string representing the main text.}

\item{motif}{A character string representing the pattern to search for.}
}
\value{
A vector of integers indicating the starting positions of each match (1-based index).
}
\description{
Searches for all occurrences of a pattern in a given text using a brute-force approach.
}
