% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{naive_pattern_matching_Rcpp}
\alias{naive_pattern_matching_Rcpp}
\title{Naive pattern matching algorithm using C++}
\usage{
naive_pattern_matching_Rcpp(texte, motif)
}
\arguments{
\item{texte}{A character string representing the main text.}

\item{motif}{A character string representing the pattern to search for.}
}
\value{
A vector of integers indicating the starting positions of each match (1-based index).
}
\description{
Searches for all occurrences of a pattern in a given text using a brute-force approach.
}
